+++ src/Rule/DateTimeRule.php
@@ -14,11 +14,15 @@ use function trim;
 class DateTimeRule extends AbstractRule
 {
     /**
+     * @param null|string $minValue The minimum value a date time string can be.
+     * @param null|string $maxValue The maximum value a date time string can be.
      * @param bool $allowNull When true, null vlaues are valid.
      * @param bool $allowEmpty When true, empty values are valid.
      * @param string $empty The value to use as an empty value.
      */
     public function __construct(
+        private ?string $minValue = null,
+        private ?string $maxValue = null,
         bool $allowNull = false,
         bool $allowEmpty = false,
         string $empty = '0000-00-00 00:00:00',
@@ -63,5 +67,56 @@ class DateTimeRule extends AbstractRule
         }
 
         return parent::clean($value);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function isValidConstraint(mixed $value): bool
+    {
+        $value = trim(strval($value));
+
+        if ($this->minValue !== null &&
+            $this->compareDateTimes($this->minValue, $value) > 0
+        ) {
+            return false;
+        }
+
+        if ($this->maxValue !== null &&
+            $this->compareDateTimes($this->maxValue, $value) < 0
+        ) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function cleanConstraint(mixed $value): mixed
+    {
+        $value = parent::cleanConstraint($value);
+
+        if (is_string($value)) {
+            if ($this->minValue !== null &&
+                $this->compareDateTimes($this->minValue, $value) > 0
+            ) {
+                return $this->minValue;
+            }
+
+            if ($this->maxValue !== null &&
+                $this->compareDateTimes($this->maxValue, $value) < 0
+            ) {
+                return $this->maxValue;
+            }
+        }
+
+        return $value;
+    }
+
+    private function compareDateTimes(string $a, string $b): int
+    {
+        return $a <=> $b;
     }
 }