+++ src/Rule/NullifyRule.php
@@ -48,6 +48,22 @@ class NullifyRule implements RuleInterface
     /**
      * @inheritdoc
      */
+    public function isValidAndClean(mixed $value): bool
+    {
+        if (!$this->isValid($value)) {
+            return false;
+        }
+
+        if ($this->clean($value) !== $value) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function getError(): ?string
     {
         return null;