+++ src/Rule/RuleInterface.php
@@ -32,6 +32,16 @@ interface RuleInterface
     public function clean(mixed $value): mixed;
 
     /**
+     * Determines if $value meets the requirements of this rule and has been
+     * cleaned.
+     *
+     * @param mixed $value The value to test.
+     * @return bool True if the value meets the requirements and is clean,
+     *  otherwise false.
+     */
+    public function isValidAndClean(mixed $value): bool;
+
+    /**
      * Gets the error name of this rule.
      *
      * @return null|string The error name of this rule if any.